+++ packages/design-system/src/interfaces/promo.d.ts
@@ -1,0 +1,18 @@
+// <reference types="react" />
+
+export type PromoSize = "small" | "medium" | "large";
+
+export interface IPromoProps extends React.HTMLAttributes<HTMLDivElement> {
+    buttonText?: string;
+    children: React.ReactNode;
+    header: string;
+    img?: string;
+    light?:boolean;
+    linkText?: string;
+    onButtonClick?: React.MouseEventHandler<HTMLButtonElement>;
+    onClose?: React.MouseEventHandler<HTMLButtonElement>;
+    onLinkClick?: React.MouseEventHandler<HTMLButtonElement>;
+    size?: PromoSize;
+}
+
+export var Promo: React.ComponentType<IPromoProps>;